o/snapstate: replace test
