+++ overlord/snapstate/autorefresh_test.go
@@ -46,7 +46,6 @@ import (
 	"github.com/snapcore/snapd/overlord/state"
 	"github.com/snapcore/snapd/release"
 	"github.com/snapcore/snapd/snap"
-	"github.com/snapcore/snapd/snap/snaptest"
 	"github.com/snapcore/snapd/store"
 	"github.com/snapcore/snapd/store/storetest"
 	"github.com/snapcore/snapd/testutil"
@@ -1626,114 +1625,4 @@ func (s *autoRefreshTestSuite) TestMaybeAsyncPendingRefreshNotificationSkips(c *
 	defer restore()
 
 	snapstate.MaybeAsyncPendingRefreshNotification(context.TODO(), s.state, &userclient.PendingSnapRefreshInfo{})
-}
-
-func (s *autoRefreshTestSuite) TestAutoRefreshSplitRefresh(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	restore := release.MockOnClassic(true)
-	defer restore()
-	restore = snapstatetest.MockDeviceModel(ModelWithBase("core18"))
-	defer restore()
-
-	snaps := []string{"snapd", "kernel", "core18", "gadget", "some-base", "some-base-snap", "some-snap-with-core18-base"}
-
-	types := map[string]string{
-		"snapd":                      "snapd",
-		"core18":                     "base",
-		"gadget":                     "gadget",
-		"kernel":                     "kernel",
-		"some-base":                  "base",
-		"some-base-snap":             "app",
-		"some-snap-with-core18-base": "app",
-	}
-	snapIds := map[string]string{
-		"snapd":                      "snapd-snap-id",
-		"core18":                     "core18-snap-id",
-		"gadget":                     "gadget-core18-id",
-		"kernel":                     "kernel-id",
-		"some-base":                  "some-base-id",
-		"some-snap-with-core18-base": "some-snap-with-core18-base",
-		"some-base-snap":             "some-base-snap-id",
-	}
-	bases := map[string]string{
-		// create an implicit dependency between the two task sets
-		"some-snap-with-core18-base": "core18",
-		"some-base-snap":             "some-base",
-	}
-
-	var paths []string
-	for _, sn := range snaps {
-		yaml := fmt.Sprintf("name: %s\nversion: 1.0\nepoch: 1\ntype: %s", sn, types[sn])
-		if base, ok := bases[sn]; ok {
-			yaml += fmt.Sprintf("\nbase: %s", base)
-		}
-
-		oldSi := &snap.SideInfo{RealName: sn, SnapID: snapIds[sn], Revision: snap.R(1)}
-		newSi := &snap.SideInfo{RealName: sn, SnapID: snapIds[sn], Revision: snap.R(11)}
-
-		path, info := snaptest.MakeTestSnapInfoWithFiles(c, yaml, nil, newSi)
-		paths = append(paths, path)
-
-		snaptest.MockSnap(c, yaml, oldSi)
-		snapstate.Set(s.state, sn, &snapstate.SnapState{
-			Active:          true,
-			Sequence:        snapstatetest.NewSequenceFromSnapSideInfos([]*snap.SideInfo{oldSi}),
-			Current:         oldSi.Revision,
-			TrackingChannel: "latest/stable",
-			SnapType:        types[sn],
-		})
-
-		s.store.refreshable = append(s.store.refreshable, info)
-	}
-
-	af := snapstate.NewAutoRefresh(s.state)
-	s.state.Unlock()
-	err := af.Ensure()
-	s.state.Lock()
-	c.Check(err, IsNil)
-
-	chgs := s.state.Changes()
-	c.Assert(chgs, HasLen, 1)
-	chg := chgs[0]
-
-	// don't start with snapd because everything depends on it
-	t := findTaskForSnap(c, chg, "prerequisites", "core18")
-	waitChain, err := getWaitChain(t)
-	c.Assert(err, IsNil)
-	c.Assert(waitChain, testutil.DeepUnsortedMatches, []string{"core18", "some-snap-with-core18-base", "gadget", "kernel"})
-
-	t = findTaskForSnap(c, chg, "prerequisites", "some-base")
-	waitChain, err = getWaitChain(t)
-	c.Assert(err, IsNil)
-	c.Assert(waitChain, DeepEquals, []string{"some-base", "some-base-snap"})
-}
-
-func getWaitChain(t *state.Task) ([]string, error) {
-	var waitChain []string
-	set := map[string]bool{t.ID(): true}
-	queue := []*state.Task{t}
-
-	for len(queue) > 0 {
-		t := queue[0]
-		queue = queue[1:]
-
-		if t.Kind() == "prerequisites" {
-			var snapsup *snapstate.SnapSetup
-			if err := t.Get("snap-setup", &snapsup); err != nil {
-				return nil, err
-			}
-			waitChain = append(waitChain, snapsup.InstanceName())
-		}
-
-		for _, next := range t.HaltTasks() {
-			if !set[next.ID()] {
-				set[next.ID()] = true
-				queue = append(queue, next)
-			}
-		}
-	}
-
-	return waitChain, nil
 }
